requirements.txt
flask
speedtest-cli
psutil
